test(compiler): refactor i18n integration test
